refactor(types): migrate double/Double to BigDecimal and add H2 test infra
Replace all double/Double fields with BigDecimal across monetary and quantity
columns to eliminate floating-point precision loss in financial calculations and
stock deductions: Product (preco, custo), RecipeIngredient (quantidade),
IngredientMapping (fatorConversao) and all their associated DTOs and service
methods. Update ProductMapper's margin expression and OrderServiceImpl's price
reduction to operate natively on BigDecimal.

Also fix JWT decoder to use HS256 symmetric secret (NimbusJwtDecoder.withSecretKey)
instead of RS256/JWKS — aligning with Supabase's actual signing algorithm — and
expose the secret via the SUPABASE_JWT_SECRET environment variable instead of
committing the raw value.

Add H2 in-memory database (test scope) and application-test.properties so the
Spring context can start in CI without a live Supabase connection; switch
SoftLanchesApplicationTests to @ActiveProfiles("test").

+++ backend/src/main/java/com/softlanches/products/dto/CreateProductRequest.java
@@ -4,6 +4,8 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.PositiveOrZero;
 import jakarta.validation.constraints.Size;
 
+import java.math.BigDecimal;
+
 public record CreateProductRequest(
 
         @NotBlank(message = "Nome é obrigatório")
@@ -11,10 +13,10 @@ public record CreateProductRequest(
         String nome,
 
         @PositiveOrZero(message = "Preço não pode ser negativo")
-        Double preco,
+        BigDecimal preco,
 
         @PositiveOrZero(message = "Custo não pode ser negativo")
-        Double custo,
+        BigDecimal custo,
 
         Boolean ativo,
