+++ backend/src/main/java/com/softlanches/recipes/dto/CreateIngredientRequest.java
@@ -4,6 +4,8 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.PositiveOrZero;
 import jakarta.validation.constraints.Size;
 
+import java.math.BigDecimal;
+
 public record CreateIngredientRequest(
 
         @NotBlank(message = "Nome do ingrediente é obrigatório")
@@ -11,7 +13,7 @@ public record CreateIngredientRequest(
         String nome,
 
         @PositiveOrZero(message = "Quantidade não pode ser negativa")
-        double quantidade,
+        BigDecimal quantidade,
 
         @NotBlank(message = "Unidade é obrigatória")
         @Size(max = 50)
